pyproject.toml
[tool.poetry]
name = "abiturientkz"
version = "0.1.0"
description = ""
authors = ["David <devid02ru@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
alembic = "^1.13.1"
pydantic-settings = "^2.2.1"
asyncpg = "^0.29.0"
greenlet = "^3.0.3"
passlib = "^1.7.4"
aiohttp = "^3.9.5"
bcrypt = "^4.1.3"
pytz = "^2024.1"
redis = "^5.0.4"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
